Added pointsblock reader and writer tests
git-svn-id: http://uuh116.googlecode.com/svn/Nebulabrot@322 333f4dc9-c705-684f-d081-88386eedeb15

+++ tst/net/lab0/nebula/core/TestOpenClMandelbrotComputeRoutines.java
@@ -62,7 +62,7 @@ public class TestOpenClMandelbrotComputeRoutines
         }
         for (int i = 0; i < side * side; ++i)
         {
-            int iter = MandelbrotComputeRoutines.computeIterationsCountOptim2(xArray[i], yArray[i], maxIter);
+            long iter = MandelbrotComputeRoutines.computeIterationsCountOptim2(xArray[i], yArray[i], maxIter);
             Assert.assertEquals("For point (" + xArray[i] + ";" + yArray[i] + ")", iter, iterations[i]);
         }
     }
